feat: atualizando controller de tarefas com o tratamento para tarefa não encontrada

+++ src/controllers/tarefaController.ts
@@ -48,11 +48,24 @@ export const updateTarefa = async (req: Request, res: Response) => {
   const { nome } = req.body;
 
   try {
-    const tarefa = await prisma.tarefa.update({
+    // Primeiro, tenta encontrar a tarefa pelo ID
+    const tarefa = await prisma.tarefa.findUnique({
+      where: { id },
+    });
+
+    // Verifica se a tarefa existe
+    if (!tarefa) {
+      res.status(404).json({ message: "Tarefa não encontrada." });
+      return;
+    }
+
+    // Se a tarefa existir, procede com a atualização
+    const updatedTarefa = await prisma.tarefa.update({
       where: { id },
       data: { nome },
     });
-    res.status(200).json(tarefa);
+
+    res.status(200).json(updatedTarefa);
   } catch (error) {
     res.status(500).json({ message: "Erro ao atualizar tarefa." });
   }
@@ -62,9 +75,22 @@ export const deleteTarefa = async (req: Request, res: Response) => {
   const { id } = req.params;
 
   try {
+    // Primeiro, tenta encontrar a tarefa pelo ID
+    const tarefa = await prisma.tarefa.findUnique({
+      where: { id },
+    });
+
+    // Verifica se a tarefa existe
+    if (!tarefa) {
+      res.status(404).json({ message: "Tarefa não encontrada." });
+      return;
+    }
+
+    // Se a tarefa existir, procede com a deleção
     await prisma.tarefa.delete({
       where: { id },
     });
+
     res.status(204).send();
   } catch (error) {
     res.status(500).json({ message: "Erro ao excluir tarefa." });
